Install micropython libraries correctly
...since bumping the version of python and libraries the libraries has
not worked. This takes care to install them correctly and deploy them to
the target.

Since we cannot install the libs in the default path on the target
system we need to have our cwd in the libs folder. It would be good to
find the micropython correspondance to PYTHONHOME and set this
environment variable instead.